pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pageant-assistant"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "langchain>=0.3",
    "langchain-community>=0.3",
    "langchain-groq>=0.2",
    "chromadb>=0.5",
    "python-dotenv>=1.0",
    "duckduckgo-search>=6.0",
    "pydantic>=2.0",
    "streamlit>=1.38,<1.54",
    "groq>=0.9",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]
observability = ["langsmith"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
# Streamlit pages must call load_dotenv() before importing our package (settings.py
# reads env vars at import time), which unavoidably puts imports after a statement.
"apps/*.py" = ["E402"]
"apps/**/*.py" = ["E402"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["integration: requires GROQ_API_KEY (may incur API costs)"]
